\name{labthree-package}
\alias{labthree-package}
\alias{labthree}
\docType{package}
\title{
\packageTitle{labthree}
}
\description{
\packageDescription{labthree}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{labthree}
\packageIndices{labthree}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{labthree}

Maintainer: \packageMaintainer{labthree}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
